+++ neo/d3xp/gamesys/Class.h
@@ -230,7 +230,7 @@ public:
 	bool						ProcessEvent( const idEventDef *ev, idEventArg arg1, idEventArg arg2, idEventArg arg3, idEventArg arg4, idEventArg arg5, idEventArg arg6, idEventArg arg7 );
 	bool						ProcessEvent( const idEventDef *ev, idEventArg arg1, idEventArg arg2, idEventArg arg3, idEventArg arg4, idEventArg arg5, idEventArg arg6, idEventArg arg7, idEventArg arg8 );
 
-	bool						ProcessEventArgPtr( const idEventDef *ev, int *data );
+	bool						ProcessEventArgPtr( const idEventDef *ev, intptr_t *data );
 	void						CancelEvents( const idEventDef *ev );
 
 	void						Event_Remove( void );
@@ -258,7 +258,7 @@ private:
 	static idList<idTypeInfo *>	types;
 	static idList<idTypeInfo *>	typenums;
 	static int					typeNumBits;
-	static int					memused;
+	static size_t				memused;
 	static int					numobjects;
 };
